Add support for EC2 roles (instance_profile_name)
The instance_profile_name parameter allows you to select
an EC2 role to assign to an instance.  Assigning an
instance_profile_name to a cluster should assign that role to
all of the instances in that cluster by default.

Added support for instance_profile_name to add_node* and to
Cluster constructors.  Also to security group tags, so that
this value is stored across multiple runs of starcluster.

Value is supported in the configuration file as well as
at the command line, both for the 'start' and 'addnode' commands.
The addnode command allows you to overwrite a cluster's default
instance_profile_name when adding one or more new nodes.